backend/requirements.txt
fastapi==0.109.0
mangum==0.20.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
boto3==1.34.14
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
icalendar==5.0.11
python-dateutil==2.8.2
uvicorn==0.27.0
